+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**ccd0**:
+- Support expansion of .webm videos.
+- Update archives with data from MayhemYDG fork.
+
 ### v2.9.11 
 *2014-04-03*